requirements.txt
arrow==1.2.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
h11==0.12.0
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
Jinja2==3.0.3
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.0.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.26.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
wsproto==1.0.0
zipp==3.6.0
